chore(federation): split out and rename SelectionSet::normalize implementation (#5593)
